Explain why the 'known place' is not defined.
One job spec says:

> # Secret and Credentials files
> 
> The authentication files that Google Api client needs to access private GSuite files can be located in a known place inside the root of the xml-authoring-project. Which is defined at: https://github.com/forikal-uk/xml-authoring-project

So, I have updated the README to explain that the design anticipates the files to be at a "known place" despite the fact that the "known place" is still be defined. Now that two other projects have connected to Google API we are almost in a position to define that place.